predictor-service/requirements.txt
Flask
requests
pandas
scikit-learn
python-dotenv
prometheus_client